+++ app/smartbrain_3000/web/_app/immutable/chunks/DBU3Kvxy.js
@@ -1,0 +1,1 @@
+import"./SnnMpD4Y.js";